EU decomp differences area.c through macro_special_objects.c (#572)
* EU decomp diffs

func_80262574() and print_intro_text()
func_8027FF44()
CamaraStatus, reset_camera(), init_camera()
CutsceneStarDance2_1()
set_focus_position()
func_80299DB4()
ActionCheepCheep1()
print_string_array_info()
Moving textures F3D
Some ingame_menu ideas
render_generic_char()
alloc_ia4_tex_from_i1() - non-matching

* Review comments regarding types

* matching alloc_ia4_tex_from_i1()

* render_generic_char_at_pos()

* func_802AD54C(), func_802AD5C8(), put_multi_text_string(), PrintGenericText()

* EU ingame_menu progress

put_hud_menu_char_umlaut()
PutString()
put_menu_char_umlaut()
PrintRegularText()
PutMiniString()
get_str_x_pos_from_center_scale()
Int2Str()

* Cleanup suggestions from simon

UNUSED in PrintGenericText() and PrintRegularText().
func_802D8134()

* func_eu_802AF1B8()

* EU ingame_menu progress

func_802D8690()
func_eu_802AF478()
func_eu_802AF590()
func_eu_802AF6D4()
func_eu_802AF758()

* func_802D8980() - non-matching

* cleanups

* Document ingame_menu characters, variables, functions

Naming cleanup
Update dialog char width table for EU

* EU decomp more functions

func_802D9030()
func_802D93E0()
do_cutscene_handler()
FR and DE cake end cutscene translations

* WIP

* EU decomp pause screens

func_802DA8EC()
PauseScreenCameraMenu()
PauseScreen2()
PauseScreen1()

* EU decomp finish ingame_menu.c less one non-matching

func_802DB840()
func_802DBBB0()
func_802DBE2C()
func_802DC330()
render_save_confirmation()
func_802DCBD4()

* Cleanups and ingame_menu documentation

* Review feedback: simon

Remove Camera padding
Remove PUT_MULTI_TEXT_STRING() macro
Remove xOffset define
Remove ingame_menu.h functions that are not called externally
Remove US/EU addresses comment

* EU corrections for toxbox and piranha plant

* Quate hates floats. One weird trick to remove them from your codebase!

Quate hates floats. So does everyone really. They are
an essential part of SM64 decompilation and N64 ROM hacking
in general.

There is a really neat trick to remove floats from your codebase
and I'll tell you all about that momentarily. First, let me summarize
my entire unrelated history to you.

I was born little endian. Everyone around me was also little endian,
so I didn't even consider the possibility that there were other
endiannesses. That is until the N64 was released, sporting its fancy,
shiny, new MIPS core with big endian. EXCUSE ME? What is this BIG
Endian? They tried to explain it to me.

"Oh, it's like little endian, but you know, they put their higher bytes
in their lower addresses."

THAT MAKES NO SENSE!

"You see, processor have a choice and in MIPS on the N64, they chose
big endian. It makes no difference to the processor. The circuitry is
wired one or the other and the processor isn't even aware."

I still didn't get it.

Then I found out N64 was made of 64-bits. That is something I couldn't
even comprehend. When I looked at myself, I felt so small and I was
afraid that I would never understand.

My world was upside down.

Oh yeah, that weird trick. Um, use typedefs in a common header.

* EU: move func_802D8980 to non-matching

* envfx_snow.c with one non-matching append_snowflake_vertex_buffer()

* EU decompile envfx_bubbles.

envfx_update_whirlpool()
envfx_bubbles_update_switch()
non-matching append_bubble_vertex_buffer()

* is_in_mystery_snow_area() is static. Thanks simon!

* macro_special_objects. Thanks M14!

* Code review: red

* s/gIngameLanguage/gInGameLanguage/g
* red wants to keep reserved identifier UB

* Code review: nim

* Better casting of sDebugSysCursor
* Fix dialog naming
* Clean up get_str_x_pos_from_center #ifdef and add comments
* replace a some `//!` with `@bug`

+++ src/game/behaviors/tox_box.inc.c
@@ -13,7 +13,7 @@ void func_802BCF78(void)
 
 void func_802BCFC4(f32 a0,f32 a1,s16 a2,s16 a3) // 0x18 0x1c 0x22 0x26
 {
-    o->oPosY = 99.41124 * sins((o->oTimer+1)/8.0f * 0x8000) + o->oHomeY + 3.0f;
+    o->oPosY = 99.41124 * sins((f32)(o->oTimer+1)/8 * 0x8000) + o->oHomeY + 3.0f;
     o->oForwardVel = a0;
     o->oUnkC0 = a1;
     o->oFaceAnglePitch += a2;
